분석요청 게시판 추가, ë 업로드된 동영상 처리 시작 가능
