+++ src/main/java/com/lyj/proj/oneteamsaproj/utils/MultipartInputStreamFileResource.java
@@ -1,0 +1,25 @@
+package com.lyj.proj.oneteamsaproj.utils;
+
+import org.springframework.core.io.InputStreamResource;
+
+import java.io.IOException;
+import java.io.InputStream;
+
+public class MultipartInputStreamFileResource extends InputStreamResource {
+    private final String filename;
+
+    public MultipartInputStreamFileResource(InputStream inputStream, String filename) {
+        super(inputStream);
+        this.filename = filename;
+    }
+
+    @Override
+    public String getFilename() {
+        return this.filename;
+    }
+
+    @Override
+    public long contentLength() throws IOException {
+        return -1; // 파일의 크기를 알 수 없도록 처리
+    }
+}
