+++ src/main/resources/application.yml
@@ -1,5 +1,8 @@
 server:
   port: 8088
+  servlet:
+    session:
+      persistent: false
 spring:
   security:
     user:
